Feature/bk rubio/fix post order ticket (#80)
* fix: resolve merge conflict in orderController and EventPlan

* feat(events): 新增活動留言與收藏功能

- 新增主辦方可查詢自己主辦活動的評論
- 新增會員活動評論 (POST/GET /api/v1/events/:eventId/comments)
- 新增會員我的最愛功能 (POST/DELETE/GET /api/v1/events/:eventId/favorites)

* feat: 新增swagger ui for MemberOrder/OrderTicket,getMemberOrder 新增status

* fix: fix拉出orderticket sendmail 到 emailUtils
